pyproject.toml
[tool.poetry]
name = "cakto-split"
version = "0.1.0"
description = "Mini split engine — Cakto backend challenge"
authors = ["Rafael"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "^5.1"
djangorestframework = "^3.15"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
pytest-cov = "^6.0"
ruff = "^0.7"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "Q"]
exclude = ["app/migrations/*"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
